+++ config.h
@@ -198,7 +198,8 @@ enum which_color_theme {
 enum shm_bit_depth {
     SHM_BITS_AUTO,
     SHM_BITS_8,
-    SHM_BITS_10
+    SHM_BITS_10,
+    SHM_BITS_16,
 };
 
 struct config {